services

+++ src/main/java/com/example/pidev/Repositories/AssuranceRepository.java
@@ -3,5 +3,5 @@ package com.example.pidev.Repositories;
 import com.example.pidev.Entities.Assurance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssuranceRepository extends JpaRepository<Assurance,Long> {
+public interface AssuranceRepository extends JpaRepository<Assurance,Integer> {
 }
